python-server/requirements.txt
# Requirements for the Python server
Flask==3.1.2
joblib==1.5.3
numpy>=1.26,<2.3
Pillow==12.1.0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.7.2
scipy>=1.11,<1.14
tensorflow==2.20.0
